[skhrapov] Silence warning about dtype in Series

See merge request gdsa/python-libs/hcl-model!10

+++ src/hcl_model/model_hcl_generic.py
@@ -77,7 +77,7 @@ class HandCraftedLinearModel(TimeSeriesModelArchetype):
         # statsmodels object with model fit results
         self._fit_results = None
         # empty Series
-        self._endog = pd.Series()
+        self._endog = pd.Series(dtype=float)
         self._exog = pd.DataFrame()
 
     def fit(
